Update accept_mod.py
should save cutoff values for splits in a "cutoff_list" dataset in the output 'jk_data_.h5' file
Note that this is kind of crude, in that cutoff values are stored manually into the array 'cutoff_list' (during the 'implement_split' function). So, one has to be careful that cutoff values are stored in the right order if one adds/removes splits from 'accept_mod.py'. I guess this is OK if the splits stay pretty constant, but might be better to automate this if the number of splits changes often; improvements are welcome! ;)